requirements-linux.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
elastic-transport==9.2.0
elasticsearch==9.2.0
exceptiongroup==1.3.1
fonttools==4.53.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.28.1
idna==3.7
jiter==0.12.0
kiwisolver==1.4.5
matplotlib==3.9.2
multidict==6.1.0
numpy==2.2.2
openai==2.8.1
packaging==24.2
pandas==2.2.2
pillow==10.4.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2024.2
urllib3==2.2.2
yarl==1.8.2